Complete GreatBuildings handler extraction
